+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Inherit `namespace` value from parent `impulse_form_with`
+
 ## [0.4.0] - 2024-03-05
 
 ### Changed